wrap server in windows
