+++ .gitignore
@@ -1,3 +1,4 @@
 bin
 temp
 .vs
+*.vc*
